improve refactor
