+++ hcl/hcl_cfg_run_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/madlambda/spells/assert"
 	"github.com/mineiros-io/terramate/errors"
 	"github.com/mineiros-io/terramate/hcl"
+	"github.com/mineiros-io/terramate/hcl/ast"
 )
 
 func TestHCLParserConfigRun(t *testing.T) {
@@ -43,10 +44,10 @@ func TestHCLParserConfigRun(t *testing.T) {
 		}
 
 		body := res.Body.(*hclsyntax.Body)
-		attrs := make(hcl.Attributes)
+		attrs := make(ast.Attributes)
 
 		for name, attr := range body.Attributes {
-			attrs[name] = hcl.NewAttribute(filepath, attr)
+			attrs[name] = ast.NewAttribute(filepath, attr)
 		}
 
 		return hcl.Config{
@@ -437,7 +438,7 @@ func TestHCLParserConfigRun(t *testing.T) {
 				},
 			},
 			want: want{
-				errs: []error{errors.E(hcl.ErrHCLSyntax,
+				errs: []error{errors.E(hcl.ErrTerramateSchema,
 					mkrange("cfg.tm", start(9, 15, 147), end(9, 21, 153)))},
 			},
 		},
@@ -489,10 +490,10 @@ func TestHCLParserConfigRun(t *testing.T) {
 			},
 			want: want{
 				errs: []error{
-					errors.E(hcl.ErrHCLSyntax,
+					errors.E(hcl.ErrTerramateSchema,
 						mkrange("cfg2.tm", start(6, 15, 84), end(6, 21, 90)),
 					),
-					errors.E(hcl.ErrHCLSyntax,
+					errors.E(hcl.ErrTerramateSchema,
 						mkrange("cfg3.tm", start(6, 15, 84), end(6, 21, 90)),
 					),
 				},
